api/requirements.txt
# =============================================================================
# api/requirements.txt - APIサーバーの依存ライブラリ
# =============================================================================
# 【このファイルの役割】
# Pythonで使うライブラリ（パッケージ）の一覧
# pip install -r requirements.txt で一括インストールできる
# =============================================================================

# ---------------------------------------------------------------------------
# Webフレームワーク
# ---------------------------------------------------------------------------
# 【FastAPI】APIサーバーを作るためのフレームワーク
# 【なぜ必要か】HTTPリクエストを受け取り、処理して返す仕組み
fastapi==0.115.6

# 【uvicorn】FastAPIを動かすためのWebサーバー
# 【なぜ必要か】FastAPI単体では動かない。サーバーが必要
# 【standardオプション】websocket等の追加機能を含む
uvicorn[standard]==0.34.0

# ---------------------------------------------------------------------------
# データベース
# ---------------------------------------------------------------------------
# 【motor】MongoDBをPythonから非同期で操作するライブラリ
# 【なぜmotorか】
#   - pymongo（同期版）もあるが、FastAPIは非同期なのでmotorが相性良い
#   - 非同期 = 重い処理を待っている間に他の処理ができる
motor==3.6.0

# ---------------------------------------------------------------------------
# データ検証
# ---------------------------------------------------------------------------
# 【pydantic】データの型チェック・バリデーションを行う
# 【なぜ必要か】APIで受け取るデータが正しい形式か自動チェック
# 【注意】agent-frameworkが2.12以上を要求するため、バージョンを上げる
pydantic>=2.12.0

# 【pydantic-settings】環境変数から設定を読み込む
# 【なぜ必要か】各ファイルでos.getenv()を書く代わりに、一箇所で管理
# 【メリット】型安全、バリデーション、.envファイル対応
pydantic-settings>=2.0.0

# ---------------------------------------------------------------------------
# ファイル操作
# ---------------------------------------------------------------------------
# 【python-multipart】ファイルアップロードを処理するために必要
# 【なぜ必要か】PDFアップロード機能で使う
python-multipart==0.0.20

# ---------------------------------------------------------------------------
# HTTP通信
# ---------------------------------------------------------------------------
# 【httpx】非同期HTTPクライアント
# 【なぜ必要か】APIからMCPサーバーに非同期でリクエストを送る
# 【requestsとの違い】requestsは同期版、httpxは非同期対応
httpx>=0.25.0

# ---------------------------------------------------------------------------
# PDF処理
# ---------------------------------------------------------------------------
# 【pdf2image】PDFを画像に変換するライブラリ
# 【なぜ必要か】GPT-4oでPDFを解析するため、先に画像に変換する
pdf2image>=1.16.0

# 【Pillow】画像処理ライブラリ
# 【なぜ必要か】画像のリサイズ、Base64変換など
Pillow>=10.0.0

# ---------------------------------------------------------------------------
# AI連携
# ---------------------------------------------------------------------------
# 【openai】OpenAI APIクライアント
# 【なぜ必要か】GPT-4oで画像から構造化データを抽出
openai>=1.0.0

# ---------------------------------------------------------------------------
# Microsoft Agent Framework
# ---------------------------------------------------------------------------
# 【agent-framework】AIエージェント構築フレームワーク
# 【なぜ必要か】
#   - Function Callingの定義が簡潔になる（デコレータで定義）
#   - OpenAI/Azure OpenAIを切り替え可能
#   - 複数エージェントの連携が容易
# 【他の選択肢】
#   - LangChain: より高機能だが複雑。今回はシンプルなAgent Frameworkを選択
#   - 直接OpenAI SDK: 現状の実装。Agent Frameworkはこれをラップ
# 【注意】バージョンを固定しないとpipが全バージョンを試して時間がかかる
agent-framework==1.0.0b260116

# 【azure-identity】Azureの認証ライブラリ
# 【なぜ必要か】Azure OpenAIを使用する際の認証（APIキー以外にも対応）
# 【注意】Azure使用時のみ必要だが、将来の切り替えに備えてインストール
azure-identity>=1.15.0
